refactor(JavaScript): move RefinementSession into its own module
