+++ Public/Scripts/Common.js
@@ -9,3 +9,15 @@ export const preventFormSubmit = (id, submitFunction) => {
         })
     })
 }
+
+export const ids = () => {
+    const url = new URL(window.location.href)
+    const paths = url.pathname.split("/")
+    const refinementSessionId = paths[paths.indexOf("refinement_sessions") + 1]
+    const userStoryId = paths[paths.indexOf("user_stories") + 1]
+
+    return {
+        refinementSessionId,
+        userStoryId,
+    }
+}
