+++ Public/script.js
@@ -1,54 +1,3 @@
-const setRefinementSessionDateToNow = () => {
-    const formattedDate = new Date().toJSON().slice(0, 10)
-    document.getElementById("date").setAttribute("value", formattedDate)
-}
-
-const removeRefinementSession = (refinementSessionId) =>
-    fetch(`refinement_sessions/${refinementSessionId}`, { method: "DELETE" })
-        .then(() => location.reload())
-
-const createRefinementSession = () => {
-    const name = document.getElementById("name").value
-    const date = document.getElementById("date").value
-    fetch("refinement_sessions", {
-        method: "POST",
-        headers: {
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ name, date }),
-    })
-    .then(() => location.reload())
-}
-
-const ids = () => {
-    const url = new URL(window.location.href)
-    const paths = url.pathname.split("/")
-    const refinementSessionId = paths[paths.indexOf("refinement_sessions") + 1]
-    const userStoryId = paths[paths.indexOf("user_stories") + 1]
-
-    return {
-        refinementSessionId,
-        userStoryId,
-    }
-}
-
-const createUserStory = () => {
-    const name = document.getElementById('name').value
-    const { refinementSessionId } = ids()
-    fetch(`${refinementSessionId}/user_stories`, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ name }),
-    })
-    .then(() => window.location.reload())
-}
-
-const removeUserStory = (userStoryId) =>
-    fetch(`${ids().refinementSessionId}/user_stories/${userStoryId}`, { method: 'DELETE' })
-        .then(() => window.location.reload())
-
 // TODO: Extract this part into its own file (See https://github.com/renaudjenny/steamScrum/issues/26)
 let webSocket = null
 const isWebSocketReady = () => {
